+++ src/ServiceMap.php
@@ -15,7 +15,6 @@ use yii\web\Application;
 use function array_key_exists;
 use function define;
 use function defined;
-use function file_exists;
 use function gettype;
 use function is_array;
 use function is_file;
@@ -24,6 +23,7 @@ use function is_readable;
 use function is_string;
 use function is_subclass_of;
 use function pathinfo;
+use function realpath;
 use function sprintf;
 
 use const PATHINFO_EXTENSION;
@@ -97,18 +97,19 @@ final class ServiceMap
      */
     public function __construct(string $configPath = '')
     {
-        if (
-            $configPath !== ''
-            && (
-                file_exists($configPath) === false
-                || is_file($configPath) === false
-                || is_readable($configPath) === false
-                || pathinfo($configPath, PATHINFO_EXTENSION) !== 'php'
-            )
-        ) {
-            throw new InvalidArgumentException(
-                sprintf('Provided config path \'%s\' must be a readable PHP file.', $configPath),
-            );
+        if ($configPath !== '') {
+            $resolvedPath = realpath($configPath);
+
+            if (
+                $resolvedPath === false
+                || is_file($resolvedPath) === false
+                || is_readable($resolvedPath) === false
+                || pathinfo($resolvedPath, PATHINFO_EXTENSION) !== 'php'
+            ) {
+                throw new InvalidArgumentException(
+                    sprintf('Provided config path \'%s\' must be a readable PHP file.', $configPath),
+                );
+            }
         }
 
         defined('YII_DEBUG') || define('YII_DEBUG', true);